+++ client/components/Chat.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import axios from 'axios'
 import {Link} from 'react-router-dom'
 import socketIOClient from 'socket.io-client'
-import Video from './Video'
+
 const location = `${window.location.hostname}:8080`
 const socket = socketIOClient() //http://127.0.0.1:8080'
 import {Picker} from 'emoji-mart'
