+++ client/components/index.js
@@ -23,4 +23,4 @@ export {default as Grades} from './Grades'
 export {default as Verification} from './Verification'
 export {default as OneStudentGrades} from './OneStudentGrades'
 export {default as TeacherGrades} from './TeacherGrades'
-export {default as Video2} from './Video2'
+export {default as Video2} from './video/Video2'
