+++ client/routes.js
@@ -19,8 +19,7 @@ import {
   Grades,
   OneStudentGrades,
   TeacherGrades,
-  Verification,
-  Video2
+  Verification
 } from './components'
 import {
   me,
@@ -33,6 +32,7 @@ import {
   getUserCourses,
   readAssignments
 } from './store'
+import Video2 from './components/video/Video2'
 class Routes extends Component {
   componentDidMount() {
     this.props.loadInitialData()
@@ -91,7 +91,6 @@ const mapState = ({user}) => {
 const mapDispatch = dispatch => {
   return {
     loadInitialData() {
-      dispatch(me())
       dispatch(getCourses())
       dispatch(getTeachers())
       dispatch(getStudents())
@@ -100,6 +99,7 @@ const mapDispatch = dispatch => {
       dispatch(readUserassignments())
       dispatch(getUserCourses())
       dispatch(readAssignments())
+      dispatch(me())
     }
   }
 }
